Support asserting for component addresses and package addresses of components from the package metadata.
