+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "common"
-version = "1.2.1"
+version = "1.2.2"
 edition = "2021"
 license = "MIT"
 description = "Common datatypes which are used in the Ociswap core pools"
