+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod math;
+pub mod metadata;
 pub mod pools;
 pub mod time;
 pub mod utils;
